fix faker usage
